requirements.txt
asgiref==3.8.1
black==24.4.2
click==8.1.7
Django==4.2.14
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
PyJWT==2.8.0
sqlparse==0.5.0
tomli==2.0.1
typing_extensions==4.12.2
